Question 24, Move Grid to its own file to avoid cyclic dep
Also implement its access method. Lacks Test
